Changement de l'organisation avec la mise en place de différents modules
