+++ src/components/resource.rs
@@ -1,0 +1,33 @@
+use bevy::prelude::*;
+
+#[derive(Resource, Debug)]
+pub struct AffichageCasesNonDecouvertes(pub bool);
+
+#[derive(Resource, Debug)]
+pub struct SeedResource {
+    pub seed: Option<u32>,
+}
+
+#[derive(Resource, Debug)]
+pub struct Compteur {
+   pub minerai: u32,
+   pub energie: u32
+}
+
+#[derive(Resource, Debug)]
+pub struct VitesseGlobale {
+    pub vitesse: u32
+}
+
+#[derive(Resource, Debug)]
+pub struct CompteurRobotsSpawn {
+    pub nombre: u32
+}
+
+#[derive(Component, Debug, PartialEq)]
+pub enum Ressource {
+    Energie,
+    Mineral,
+    LieuInteretScientifique,
+    Obstacle,
+}
