+++ src/components/ui.rs
@@ -1,0 +1,10 @@
+use bevy::prelude::*;
+
+#[derive(Component)]
+pub struct TexteEnergie;
+
+#[derive(Component)]
+pub struct TexteMinerai;
+
+#[derive(Component)]
+pub struct TexteVitesse;
